requirements.txt
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.4
distro==1.5.0
Django==3.1.6
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.28.2
dockerpty==0.4.1
docopt==0.6.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
paramiko==2.7.2
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.26.3
websocket-client==0.57.0
